Fix setting ValidForGroups on MyCameraBlock View action (#2)
Using `CustomActionGetter` to flip `ValidForGroups` wasn't reliable. It
would work for a time, then stop working for a given world. Also worth
noting that `CustomActionGetter` runs every tick as long as you're in a
control seat, so it's probably not desirable to do much there at all,
especially seeing as we're not adding any new actions.